+++ package.json
@@ -13,6 +13,7 @@
     "@tanstack/react-query": "^5.59.0",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-paginate": "^8.2.0",
     "react-router-dom": "^6.26.2"
   },
   "devDependencies": {
