+++ src/entities/items/hooks/useResponsiveDisplayCount.jsx
